Nuke openssl_init.h
