+++ src/crypto/openssl/hash.cpp
@@ -3,8 +3,6 @@
 
 #include "crypto/openssl/hash.h"
 
-#include "ccf/crypto/openssl_init.h"
-
 #include <limits>
 #include <openssl/evp.h>
 #include <openssl/sha.h>
@@ -53,92 +51,13 @@ namespace ccf::crypto
     }
   }
 
-  using namespace OpenSSL;
-
-  namespace
+  void openssl_sha256(const std::span<const uint8_t>& data, uint8_t* h)
   {
     thread_local EVP_MD_CTX* mdctx = nullptr;
     thread_local EVP_MD_CTX* basectx = nullptr;
-  }
 
-  void openssl_sha256_init()
-  {
-    if (mdctx != nullptr || basectx != nullptr)
-    {
-      return; // Already initialised
-    }
-
-    mdctx = EVP_MD_CTX_new();
-    if (mdctx == nullptr)
-    {
-      throw std::logic_error("openssl_sha256_init: failed to create mdctx");
-    }
-    basectx = EVP_MD_CTX_new();
-    if (basectx == nullptr)
-    {
-      mdctx = nullptr;
-      throw std::logic_error("openssl_sha256_init: failed to create basectx");
-    }
-    if (EVP_DigestInit_ex(basectx, EVP_sha256(), nullptr) != 1)
-    {
-      mdctx = nullptr;
-      basectx = nullptr;
-      throw std::logic_error("EVP_DigestInit_ex failed");
-    }
-  }
-
-  void openssl_sha256_shutdown()
-  {
-    if (mdctx != nullptr)
-    {
-      EVP_MD_CTX_free(mdctx);
-      mdctx = nullptr;
-    }
-    if (basectx != nullptr)
-    {
-      EVP_MD_CTX_free(basectx);
-      basectx = nullptr;
-    }
-  }
-
-  void openssl_sha256(const std::span<const uint8_t>& data, uint8_t* h)
-  {
-    // EVP_Digest calls are notoriously slow with OpenSSL 3.x (see
-    // https://github.com/openssl/openssl/issues/19612). Instead, we skip the
-    // calls to EVP_DigestInit_ex() by keeping 2 static thread-local contexts
-    // and reusing them between calls. This is about 2x faster than EVP_Digest
-    // for 128-byte buffers.
-
-    if (mdctx == nullptr || basectx == nullptr)
-    {
-      throw std::logic_error(
-        "openssl_sha256 failed: openssl_sha256_init should be called first");
-    }
-
-    int rc = EVP_MD_CTX_copy_ex(mdctx, basectx);
-    if (rc != 1)
-    {
-      throw std::logic_error(fmt::format("EVP_MD_CTX_copy_ex failed: {}", rc));
-    }
-    rc = EVP_DigestUpdate(mdctx, data.data(), data.size());
-    if (rc != 1)
-    {
-      throw std::logic_error(fmt::format("EVP_DigestUpdate failed: {}", rc));
-    }
-    rc = EVP_DigestFinal_ex(mdctx, h, nullptr);
-    if (rc != 1)
-    {
-      throw std::logic_error(fmt::format("EVP_DigestFinal_ex failed: {}", rc));
-    }
-  }
-
-  void openssl_sha256_(const std::span<const uint8_t>& data, uint8_t* h)
-  {
     struct Sha256Context
     {
-      EVP_MD_CTX* mdctx = nullptr;
-      EVP_MD_CTX* basectx = nullptr;
-
       Sha256Context()
       {
         mdctx = EVP_MD_CTX_new();
@@ -177,17 +96,17 @@ namespace ccf::crypto
 
     thread_local Sha256Context ctx;
 
-    int rc = EVP_MD_CTX_copy_ex(ctx.mdctx, ctx.basectx);
+    int rc = EVP_MD_CTX_copy_ex(mdctx, basectx);
     if (rc != 1)
     {
       throw std::logic_error(fmt::format("EVP_MD_CTX_copy_ex failed: {}", rc));
     }
-    rc = EVP_DigestUpdate(ctx.mdctx, data.data(), data.size());
+    rc = EVP_DigestUpdate(mdctx, data.data(), data.size());
     if (rc != 1)
     {
       throw std::logic_error(fmt::format("EVP_DigestUpdate failed: {}", rc));
     }
-    rc = EVP_DigestFinal_ex(ctx.mdctx, h, nullptr);
+    rc = EVP_DigestFinal_ex(mdctx, h, nullptr);
     if (rc != 1)
     {
       throw std::logic_error(fmt::format("EVP_DigestFinal_ex failed: {}", rc));
