+++ src/enclave/enclave.h
@@ -2,7 +2,6 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 #include "ccf/app_interface.h"
-#include "ccf/crypto/openssl_init.h"
 #include "ccf/ds/logger.h"
 #include "ccf/js/core/context.h"
 #include "ccf/node_context.h"
@@ -94,8 +93,6 @@ namespace ccf
       rpc_map(std::make_shared<RPCMap>()),
       rpcsessions(std::make_shared<RPCSessions>(*writer_factory, rpc_map))
     {
-      ccf::crypto::openssl_sha256_init();
-
       to_host = writer_factory->create_writer_to_outside();
 
       LOG_TRACE_FMT("Creating ledger secrets");
@@ -183,7 +180,6 @@ namespace ccf
     ~Enclave()
     {
       LOG_TRACE_FMT("Shutting down enclave");
-      ccf::crypto::openssl_sha256_shutdown();
     }
 
     CreateNodeStatus create_new_node(
@@ -237,7 +233,6 @@ namespace ccf
 
     bool run_main()
     {
-      ccf::crypto::openssl_sha256_init();
       LOG_DEBUG_FMT("Running main thread");
 
       {
@@ -421,8 +416,6 @@ namespace ccf
         LOG_INFO_FMT("Enclave stopped successfully. Stopping host...");
         RINGBUFFER_WRITE_MESSAGE(AdminMessage::stopped, to_host);
 
-        ccf::crypto::openssl_sha256_shutdown();
-
         return true;
       }
     }
@@ -439,7 +432,6 @@ namespace ccf
 
     bool run_worker()
     {
-      ccf::crypto::openssl_sha256_init();
       LOG_DEBUG_FMT("Running worker thread");
 
       {
@@ -449,7 +441,6 @@ namespace ccf
           msg->data.tid, std::move(msg));
 
         ::threading::ThreadMessaging::instance().run();
-        ccf::crypto::openssl_sha256_shutdown();
       }
 
       return true;
