+++ src/host/run.cpp
@@ -3,7 +3,6 @@
 
 #include "ccf/run.h"
 
-#include "ccf/crypto/openssl_init.h"
 #include "ccf/crypto/pem.h"
 #include "ccf/crypto/symmetric_key.h"
 #include "ccf/ds/logger.h"
@@ -111,8 +110,6 @@ namespace ccf
       LOG_FAIL_FMT("Failed to ignore SIGPIPE");
       return 1;
     }
-
-    ccf::crypto::openssl_sha256_init();
 
     CLI::App app{
       "Run a single CCF node, based on the given configuration file.\n"
@@ -1032,7 +1029,6 @@ namespace ccf
       uv_walk(uv_default_loop(), cb, nullptr);
     }
     curl_global_cleanup();
-    ccf::crypto::openssl_sha256_shutdown();
 
     return loop_close_rc;
   }
