+++ src/indexing/test/indexing.cpp
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 
-#include "ccf/crypto/openssl_init.h"
 #include "ccf/ds/x509_time_fmt.h"
 #include "ccf/indexing/strategies/seqnos_by_key_bucketed.h"
 #include "ccf/indexing/strategies/seqnos_by_key_in_memory.h"
@@ -507,7 +506,6 @@ TEST_CASE(
   std::atomic<size_t> writes_to_42 = 0;
 
   auto tx_advancer = [&]() {
-    ccf::crypto::openssl_sha256_init();
     size_t i = 0;
     while (i < 1'000)
     {
@@ -529,7 +527,7 @@ TEST_CASE(
       ++i;
       std::this_thread::yield();
     }
-    ccf::crypto::openssl_sha256_shutdown();
+
     finished = true;
   };
 
@@ -569,7 +567,6 @@ TEST_CASE(
   std::atomic<bool> work_done = false;
 
   std::thread index_ticker([&]() {
-    ccf::crypto::openssl_sha256_init();
     while (!work_done)
     {
       size_t post_work_done_loops = 0;
@@ -623,7 +620,6 @@ TEST_CASE(
         std::this_thread::yield();
       }
     }
-    ccf::crypto::openssl_sha256_shutdown();
   });
 
   std::vector<std::thread> threads;
@@ -797,7 +793,6 @@ TEST_CASE(
   std::atomic<size_t> writes_to_42 = 0;
 
   auto tx_advancer = [&]() {
-    ccf::crypto::openssl_sha256_init();
     size_t i = 0;
     constexpr auto tx_count =
 #if NDEBUG
@@ -827,12 +822,10 @@ TEST_CASE(
       std::this_thread::yield();
     }
     all_submitted = true;
-    ccf::crypto::openssl_sha256_shutdown();
   };
 
   auto get_all =
     [&](const std::string& key) -> std::optional<ccf::SeqNoCollection> {
-    ccf::crypto::openssl_sha256_init();
     const auto max_range = index_a->max_requestable_range();
     auto range_start = 0;
 
@@ -864,7 +857,6 @@ TEST_CASE(
 
       if (range_end == end_seqno)
       {
-        ccf::crypto::openssl_sha256_shutdown();
         return all_results;
       }
       else
@@ -872,7 +864,6 @@ TEST_CASE(
         range_start = range_end + 1;
       }
     }
-    ccf::crypto::openssl_sha256_shutdown();
   };
 
   auto fetch_index_a = [&]() {
@@ -929,7 +920,6 @@ TEST_CASE(
   });
 
   std::thread index_ticker([&]() {
-    ccf::crypto::openssl_sha256_init();
     while (!work_done)
     {
       while (indexer.update_strategies(step_time, kv_store.current_txid()))
@@ -937,7 +927,6 @@ TEST_CASE(
         std::this_thread::yield();
       }
     }
-    ccf::crypto::openssl_sha256_shutdown();
   });
 
   std::thread watchdog([&]() {
@@ -966,11 +955,11 @@ TEST_CASE(
 int main(int argc, char** argv)
 {
   threading::ThreadMessaging::init(1);
-  ccf::crypto::openssl_sha256_init();
+
   doctest::Context context;
   context.applyCommandLine(argc, argv);
   int res = context.run();
-  ccf::crypto::openssl_sha256_shutdown();
+
   if (context.shouldExit())
     return res;
   return res;
