+++ src/node/test/receipt.cpp
@@ -4,7 +4,6 @@
 #include "ccf/receipt.h"
 
 #include "ccf/crypto/key_pair.h"
-#include "ccf/crypto/openssl_init.h"
 #include "ccf/ds/x509_time_fmt.h"
 #include "ccf/service/tables/nodes.h"
 #include "crypto/openssl/hash.h"
@@ -171,7 +170,6 @@ TEST_CASE("JSON parsing" * doctest::test_suite("receipt"))
 
 TEST_CASE("JSON roundtrip" * doctest::test_suite("receipt"))
 {
-  ccf::crypto::openssl_sha256_init();
   {
     std::shared_ptr<ccf::Receipt> r = nullptr;
     nlohmann::json j;
@@ -196,5 +194,4 @@ TEST_CASE("JSON roundtrip" * doctest::test_suite("receipt"))
       compare_receipts(p_receipt, parsed);
     }
   }
-  ccf::crypto::openssl_sha256_shutdown();
 }
