+++ tests/perf-system/submitter/submit.cpp
@@ -4,7 +4,7 @@
 #include "ccf/crypto/verifier.h"
 #include "ccf/ds/logger.h"
 #include "ccf/service/node_info_network.h"
-#include "ccf/crypto/openssl_init.h"
+
 #include "clients/perf/perf_client.h"
 #include "clients/rpc_tls_client.h"
 #include "crypto/openssl/hash.h"
@@ -268,7 +268,7 @@ int main(int argc, char** argv)
 
   ccf::logger::config::default_init();
   ccf::logger::config::level() = ccf::LoggerLevel::INFO;
-  ccf::crypto::openssl_sha256_init();
+  
   CLI::App cli_app{"Perf Tool"};
   ArgumentParser args("Perf Tool", cli_app);
   CLI11_PARSE(cli_app, argc, argv);
@@ -382,7 +382,7 @@ int main(int argc, char** argv)
   }
 
   store_parquet_results(args, data_handler);
-  ccf::crypto::openssl_sha256_shutdown();
+  
 
   return 0;
 }
